Fix handling of GOPATH with multi-values.
When using gvm it can setup a $GOPATH as the following:

GOPATH=/apps/we/are/not/in/oz:/apps/somewhere/over/the/rainbow.gvm_local/pkgsets/go1.4/local:/apps/golang/gvm/pkgsets/go1.4/global

Gosh currently does not support this sort of string and will throw a
strange error like the following:

"2015/05/08 21:25:43 Could not determine missing dependencies for
github.com/oz/the/man/behind/the/curtain: Cannot get json for
'github.com/oz/the/main/behind/the/curtain' as it is not a go package. Original
error: EOF"

This fix should allow both single and multi values. It operates the same
as "go get" does in that it uses the first value in the list [0] when
installing the go dep.